Naredil komponento za upravljanje z varovanci
