+++ frontend/angular/src/app/components/pages/home/home.component.html
@@ -1,16 +1,5 @@
 <h3>Skavtko home page</h3>
-<button (click)="openClanPostModal()">Open Modal</button>
-
-<app-clan-post-modal #clanPostModal></app-clan-post-modal>
-<h4>Moji varovanci</h4>
-<ul>
-  <li *ngFor="let varovanec of mojiVarovanci">
-    <p>{{varovanec.ime}} {{varovanec.priimek}}</p>
-    <button (click)="openClanPutModal(varovanec)">Uredi</button>
-    <button (click)="odstraniVarovanca(varovanec.id)">Zbriši</button>
-  </li>
-</ul>
-<app-clan-put-modal #clanPutModal></app-clan-put-modal>
+<app-varovanci></app-varovanci>
 <h4>Moje skupine:</h4>
 <app-skupina-post-modal-button></app-skupina-post-modal-button>
 <p>Skavtko je aplikacija...</p>
