+++ frontend/angular/src/app/components/pages/home/home.component.ts
@@ -5,53 +5,19 @@ import { CommonModule } from '@angular/common';
 import { ClanPutModalComponent } from '../../modal/clan-put-modal/clan-put-modal.component';
 import { ClanDataService } from '../../../services/clan-data.service';
 import { SkupinaPostModalButtonComponent } from "../../modal-button/skupina-post-modal-button/skupina-post-modal-button.component";
+import { VarovanciComponent } from "../../add-in/varovanci/varovanci.component";
 
 @Component({
   selector: 'app-home',
   standalone: true,
-  imports: [ClanPostModalComponent, CommonModule, ClanPutModalComponent, SkupinaPostModalButtonComponent],
+  imports: [CommonModule, SkupinaPostModalButtonComponent, VarovanciComponent],
   templateUrl: './home.component.html',
   styleUrl: './home.component.css'
 })
 export class HomeComponent {
 
-  constructor(private userData: UserDataService, private clanData: ClanDataService) {}
-
-  mojiVarovanci: Clan[] = [];
+  constructor() {}
 
   ngOnInit(): void {
-    if(localStorage.getItem('user') && this.userData.varovanciNotInit()){
-      const user: Clan = JSON.parse(localStorage.getItem('user') as string);
-      this.clanData.getVarovanci(user.id).subscribe((response) => {
-        this.userData.initVarovanci(response);
-      })
-    }
-    this.userData.varovanci$.subscribe((varovanci) => {
-      this.mojiVarovanci = varovanci;
-    });
-  }
-
-  // Modalno okno za dodajanje clana
-  @ViewChild('clanPostModal') clanPostModal: ClanPostModalComponent | undefined;
-
-  openClanPostModal() : void {
-    if(this.clanPostModal) {
-      this.clanPostModal.open();
-    }
-  }
-  @ViewChild('clanPutModal') clanPutModal: ClanPutModalComponent | undefined;
-
-  openClanPutModal(clan : Clan): void {
-    if(this.clanPutModal) {
-      this.clanPutModal.open(clan);
-    }
-  }
-
-  odstraniVarovanca(id: number){
-    this.clanData.deleteClan(id).subscribe(
-      (response) => {
-        this.userData.odstraniVarovanca(id);
-      }
-    )
   }
 }
